[Tests] TestUtils moved
